refactor: created users app
