+++ train_api/urls.py
@@ -1,7 +1,6 @@
 from django.urls import path
 
 from .views import (
-    UserViewSet,
     CrewViewSet,
     StationViewSet,
     TrainTypeViewSet,
@@ -9,15 +8,11 @@ from .views import (
     RouteViewSet,
     JourneyViewSet,
     OrderViewSet,
-    TicketViewSet, ManageUserView,
+    TicketViewSet,
 )
 from .constants import DEFAULT_LIST_ACTIONS, DEFAULT_DETAIL_ACTIONS
 
 urlpatterns = [
-    path("users/", UserViewSet.as_view(DEFAULT_LIST_ACTIONS), name="user_list"),
-    path("users/<int:pk>", UserViewSet.as_view(DEFAULT_DETAIL_ACTIONS), name="user_detail"),
-    path("me/", ManageUserView.as_view(), name="manage_user"),
-
     path("crews/", CrewViewSet.as_view(DEFAULT_LIST_ACTIONS), name="crew_list"),
     path("crews/<int:pk>", CrewViewSet.as_view(DEFAULT_DETAIL_ACTIONS), name="crew_detail"),
