+++ train_api/views.py
@@ -1,11 +1,9 @@
-from django.contrib.auth.models import User
-from rest_framework import viewsets, status, generics
+from rest_framework import viewsets, status
 from rest_framework.decorators import action
 from rest_framework.response import Response
 
 from train_api.models import Crew, Station, TrainType, Train, Route, Journey, Order, Ticket
 from train_api.serializers import (
-    UserSerializer,
     CrewSerializer,
     StationListSerializer,
     StationDetailSerializer,
@@ -23,18 +21,6 @@ from train_api.serializers import (
     JourneyCreateSerializer,
     TrainImageSerializer,
 )
-
-
-class UserViewSet(viewsets.ModelViewSet):
-    queryset = User.objects.all().order_by("-date_joined")
-    serializer_class = UserSerializer
-
-
-class ManageUserView(generics.RetrieveUpdateAPIView):
-    serializer_class = UserSerializer
-
-    def get_object(self):
-        return self.request.user
 
 
 class CrewViewSet(viewsets.ModelViewSet):
