+++ users/constants.py
@@ -1,0 +1,11 @@
+DEFAULT_LIST_ACTIONS = {
+    "get": "list",
+    "post": "create",
+}
+
+DEFAULT_DETAIL_ACTIONS = {
+    "get": "retrieve",
+    "put": "update",
+    "patch": "partial_update",
+    "delete": "destroy"
+}
